// qto-ui/src/app/models/constants/service-type.ts
export enum ServiceType {
  BROADBAND = "Broadband",
  DIA = "DIA",
  UCAAS = "UCaaS",
  G = "4G/5G",
  CROSSCONNECT = "Cross Connect",
  ETHERNET = "Ethernet",
  TELEVISION = "Television",
  MPLS = "MPLS",
  THREATMDR = "threatMDR",
  RANSOMMDR = "ransomMDR",
  RISKMDR = "riskMDR",
  ENGINEERING_IAM="Engineering-IAM",
  ENGINEERING_MDM="Engineering-MDM",
  ENGINEERING_ENDPOINT="Engineering-Endpoint",
  CYBER360MXDR = "cyber360-MXDR",
  ENGINEERING_INFO_PROTECTION="Engineering-Info Protection",
  ENGINEERING_EMAIL_MESSAGING="Engineering-Email & Messaging",
  MICROSOFTLICENSES= "Microsoft Licenses"
}